Update bin/build_pgms.sh to just use make
There's no good reason why this script is needed to build with -DNDEBUG
and without -g. The performance of the programs is more than adequate
even with assert(std::is_sorted(...)) checks in them.

Just make this script use 'make clean pgms' so that it doesn't need to
be maintained to add new source files, so that the .o files can be
reused for multiple programs, and so that it shares the same build flags
as the makefile.